Modify Project Structure
This modifies the project structure to allow for easily addiing functions to convert Rhino objects into Blender equivlents as well as facilitates the creation of a typical blender addon.
