+++ import_3dm/__init__.py
@@ -1,0 +1,101 @@
+""" MIT License
+
+Copyright (c) 2018 Nathan Letwory , Joel Putnam
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE. """
+
+bl_info = {
+    "name": "Import Rhinoceros 3D",
+    "author": "jesterKing, Joel Putnam",
+    "version": (0, 0, 1),
+    "blender": (2, 80, 0),
+    "location": "File > Import > Rhinoceros 3D (.3dm)",
+    "description": "This addon lets you import Rhinoceros 3dm files",
+    "warning": "The importer doesn't handle all data in 3dm files yet",
+    "wiki_url": "https://github.com/jesterKing/import_3dm",
+    "category": "Import-Export",
+}
+
+import os.path
+import binascii
+import struct
+
+import bpy
+# ImportHelper is a helper class, defines filename and
+# invoke() function which calls the file selector.
+from bpy_extras.io_utils import ImportHelper
+from bpy.props import StringProperty, BoolProperty, EnumProperty
+from bpy.types import Operator
+
+import rhino3dm as r3d
+
+from .read3dm import read_3dm
+
+class Import3dm(Operator, ImportHelper):
+    """Import Rhinoceros 3D files (.3dm). Currently does render meshes only, more geometry and data to follow soon."""
+    bl_idname = "import_3dm.some_data"  # important since its how bpy.ops.import_3dm.some_data is constructed
+    bl_label = "Import Rhinoceros 3D file"
+
+    # ImportHelper mixin class uses this
+    filename_ext = ".3dm"
+
+    filter_glob: StringProperty(
+        default="*.3dm",
+        options={'HIDDEN'},
+        maxlen=1024,  # Max internal buffer length, longer would be clamped.
+    )
+
+    # List of operator properties, the attributes will be assigned
+    # to the class instance from the operator settings before calling.
+    import_hidden: BoolProperty(
+        name="Import Hidden Geometry",
+        description="Import Hidden Geometry",
+        default=True,
+    )
+
+#    type: EnumProperty(
+#        name="Example Enum",
+#        description="Choose between two items",
+#        items=(
+#            ('OPT_A', "First Option", "Description one"),
+#            ('OPT_B', "Second Option", "Description two"),
+#        ),
+#        default='OPT_A',
+#    )
+
+    def execute(self, context):
+        return read_3dm(context, self.filepath, self.import_hidden)
+
+# Only needed if you want to add into a dynamic menu
+def menu_func_import(self, context):
+    self.layout.operator(Import3dm.bl_idname, text="Rhinoceros 3D (.3dm)")
+
+def register():
+    bpy.utils.register_class(Import3dm)
+    bpy.types.TOPBAR_MT_file_import.append(menu_func_import)
+
+def unregister():
+    bpy.utils.unregister_class(Import3dm)
+    bpy.types.TOPBAR_MT_file_import.remove(menu_func_import)
+
+if __name__ == "__main__":
+    register()
+
+    # test call
+    bpy.ops.import_3dm.some_data('INVOKE_DEFAULT')
