+++ import_3dm/converters/__init__.py
@@ -1,0 +1,38 @@
+""" MIT License
+
+Copyright (c) 2018 Nathan Letwory , Joel Putnam
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE. """
+
+import rhino3dm as r3d
+
+from .material import *
+from .layers import *
+from .render_mesh import import_render_mesh
+from .curve import import_curve
+
+'''
+Dictionary mapping between the Rhino file types and importer functions
+'''
+RHINO_TYPE_TO_IMPORT = {
+    r3d.ObjectType.Brep : import_render_mesh,
+    r3d.ObjectType.Extrusion : import_render_mesh,
+    r3d.ObjectType.Mesh : import_render_mesh,
+    r3d.ObjectType.Curve : import_curve
+}
